refactor: split UI and component logic
