--- src/components/todo/list/style.scss
@@ -1,10 +1,0 @@
-ul.filters {
-  padding: 0px;
-
-  li {
-    &:not(:last-child) {
-      margin-right: 10px;
-    }
-    display: inline-block;
-  }
-}
